requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.3.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
Werkzeug==2.3.4
requests-mock==1.10.0
clyent==1.2.1
nbformat==5.4.0